split the 2 resoures into 2
