--- docs/resources/packetfabric_cloud_provider_credential.md
@@ -1,61 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "packetfabric_cloud_provider_credential Resource - packetfabric"
-subcategory: ""
-description: |-
-  
----
-
-# packetfabric_cloud_provider_credential (Resource)
-
-
-
-## Example Usage
-
-```terraform
-resource "packetfabric_cloud_provider_credential" "aws_creds_staged" {
-  provider       = packetfabric
-  cloud_provider = "aws"
-  description    = "AWS Staging Environement"
-  aws_access_key = var.pf_aws_key
-  aws_secret_key = var.pf_aws_secret
-}
-
-output "packetfabric_cloud_provider_credential" {
-  value     = packetfabric_cloud_provider_credential.aws_creds_staged
-  sensitive = true
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `cloud_provider` (String) The cloud provider of this cloud provider credential.
-
-	Enum: ["aws" or "google"]
-- `description` (String) Description of the Cloud Provider Credentials.
-
-### Optional
-
-- `aws_access_key` (String, Sensitive) The AWS access key you want to save. Can also be set with the PF_AWS_ACCESS_KEY_ID environment variable.
-- `aws_secret_key` (String, Sensitive) The AWS secret key you want to save. Can also be set with the PF_AWS_SECRET_ACCESS_KEY environment variable.
-- `service_account` (String, Sensitive) The Google service account JSON you want to save. Can also be set with the PF_GOOGLE_CREDENTIALS environment variable.
-- `timeouts` (Block, Optional) (see [below for nested schema](#nestedblock--timeouts))
-
-### Read-Only
-
-- `id` (String) The ID of this resource.
-
-<a id="nestedblock--timeouts"></a>
-### Nested Schema for `timeouts`
-
-Optional:
-
-- `create` (String)
-- `delete` (String)
-- `read` (String)
-- `update` (String)
-
-
