+++ docs/resources/packetfabric_cloud_provider_credential_aws.md
@@ -1,0 +1,45 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "packetfabric_cloud_provider_credential_aws Resource - terraform-provider-packetfabric"
+subcategory: ""
+description: |-
+  
+---
+
+# packetfabric_cloud_provider_credential_aws (Resource)
+
+Adding AWS Cloud Provider credentials to your PacketFabric environment allows you to perform certain tasks from within the PacketFabric portal.
+
+## Example Usage
+
+```terraform
+resource "packetfabric_cloud_provider_credential_aws" "aws_creds_staged" {
+  provider       = packetfabric
+  description    = "AWS Staging Environement"
+  aws_access_key = var.pf_aws_key # or use env var PF_AWS_ACCESS_KEY_ID
+  aws_secret_key = var.pf_aws_secret # or use env var PF_AWS_SECRET_ACCESS_KEY
+}
+
+output "packetfabric_cloud_provider_credential_aws" {
+  value     = packetfabric_cloud_provider_credential_aws.aws_creds_staged
+  sensitive = true
+}
+```
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `description` (String) Description of the Cloud Provider Credentials.
+
+### Optional
+
+- `aws_access_key` (String, Sensitive) The AWS access key you want to save. Can also be set with the AWS_ACCESS_KEY_ID environment variable.
+- `aws_secret_key` (String, Sensitive) The AWS secret key you want to save. Can also be set with the AWS_SECRET_ACCESS_KEY environment variable.
+
+### Read-Only
+
+- `id` (String) The ID of this resource.
+
+
