+++ docs/resources/packetfabric_cloud_provider_credential_google.md
@@ -1,0 +1,43 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "packetfabric_cloud_provider_credential_google Resource - terraform-provider-packetfabric"
+subcategory: ""
+description: |-
+  
+---
+
+# packetfabric_cloud_provider_credential_google (Resource)
+
+Adding Google Cloud Provider credentials to your PacketFabric environment allows you to perform certain tasks from within the PacketFabric portal.
+
+## Example Usage
+
+```terraform
+resource "packetfabric_cloud_provider_credential_google" "google_creds_staged" {
+  provider        = packetfabric
+  description     = "Google Staging Environement"
+  service_account = var.service_account # or use env var GOOGLE_CREDENTIALS
+}
+
+output "packetfabric_cloud_provider_credential_google" {
+  value     = packetfabric_cloud_provider_credential_google.google_creds_staged
+  sensitive = true
+}
+```
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `description` (String) Description of the Cloud Provider Credentials.
+
+### Optional
+
+- `service_account` (String, Sensitive) The Google service account JSON you want to save. Can also be set with the GOOGLE_CREDENTIALS environment variable.
+
+### Read-Only
+
+- `id` (String) The ID of this resource.
+
+
