+++ examples/resources/packetfabric_cloud_provider_credential_google/resource.tf
@@ -1,0 +1,10 @@
+resource "packetfabric_cloud_provider_credential_google" "google_creds_staged" {
+  provider        = packetfabric
+  description     = "Google Staging Environement"
+  service_account = var.service_account # or use env var GOOGLE_CREDENTIALS
+}
+
+output "packetfabric_cloud_provider_credential_google" {
+  value     = packetfabric_cloud_provider_credential_google.google_creds_staged
+  sensitive = true
+}
