+++ templates/resources/cloud_provider_credential_aws.md.tmpl
@@ -1,0 +1,17 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "packetfabric_cloud_provider_credential_aws Resource - terraform-provider-packetfabric"
+subcategory: ""
+description: |-
+  
+---
+
+# packetfabric_cloud_provider_credential_aws (Resource)
+
+Adding AWS Cloud Provider credentials to your PacketFabric environment allows you to perform certain tasks from within the PacketFabric portal.
+
+## Example Usage
+
+{{tffile "examples/resources/packetfabric_cloud_provider_credential_aws/resource.tf"}}
+
+{{ .SchemaMarkdown }}
