+++ templates/resources/cloud_provider_credential_google.md.tmpl
@@ -1,0 +1,17 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "packetfabric_cloud_provider_credential_google Resource - terraform-provider-packetfabric"
+subcategory: ""
+description: |-
+  
+---
+
+# packetfabric_cloud_provider_credential_google (Resource)
+
+Adding Google Cloud Provider credentials to your PacketFabric environment allows you to perform certain tasks from within the PacketFabric portal.
+
+## Example Usage
+
+{{tffile "examples/resources/packetfabric_cloud_provider_credential_google/resource.tf"}}
+
+{{ .SchemaMarkdown }}
