--- templates/resources/cloud_provider_credentials.tmpl
@@ -1,18 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "packetfabric_cloud_provider_credential Resource - terraform-provider-packetfabric"
-subcategory: ""
-description: |-
-  
----
-
-# packetfabric_cloud_provider_credential (Resource)
-
-Adding Cloud Providers credentials to your PacketFabric environment allows you to perform certain tasks from within the PacketFabric portal.
-
-
-## Example Usage
-
-{{tffile "examples/resources/packetfabric_cloud_provider_credential/resource.tf"}}
-
-{{ .SchemaMarkdown }}
